+++ tests/integration/prepare.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+set -euo pipefail
+
+TESTING_MODEL="$(juju switch)"
+
+# lxd should be install and init by a previous step in integration test action.
+echo "bootstrapping lxd juju controller"
+juju bootstrap localhost localhost
+
+echo "Switching to testing model"
+juju switch "$TESTING_MODEL"